Simplify result data type of convertToArabicNumber: Integer => int.
Note: this is Nelkinda feedback #29.

+++ 008_RomanNumerals/src/test/java/de/kifaru/ndegendogo/kata/romanNumbers/RomanNumbersConverterTestGoodcase.java
@@ -52,8 +52,8 @@ public class RomanNumbersConverterTestGoodcase {
     @Test
     public void testConvertToArabicNumber() {
         try {
-            final Integer actualResult = new RomanNumbersConverter().convertToArabicNumber(actualInput);
-            assertEquals(expectedResult, actualResult.intValue());
+            final int actualResult = new RomanNumbersConverter().convertToArabicNumber(actualInput);
+            assertEquals(expectedResult, actualResult);
         } catch (RomanNumberFormatException e) {
             fail();
         }
